+++ pkg/store/lani/decode_test.go
@@ -1,4 +1,4 @@
-package store_test
+package lani_test
 
 import (
 	"crypto/rand"
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/oklog/ulid/v2"
-	. "github.com/rotationalio/honu/pkg/store"
+	. "github.com/rotationalio/honu/pkg/store/lani"
 	"github.com/stretchr/testify/require"
 )
 
@@ -344,12 +344,12 @@ func TestDecoder(t *testing.T) {
 
 	t.Run("Struct", func(t *testing.T) {
 		t.Run("Nil", func(t *testing.T) {
-			var obj *Compression
+			var obj *Mock
 
 			_, err := enc.EncodeStruct(obj)
 			require.NoError(t, err, "could not encode nil struct input")
 
-			var cmp *Compression
+			var cmp *Mock
 			isNil, err := NewDecoder(enc.Bytes()).DecodeStruct(cmp)
 			require.NoError(t, err, "could not decode nil struct")
 			require.True(t, isNil, "expected isNil to be true")
