+++ pkg/store/lani/errors.go
@@ -1,4 +1,4 @@
-package store
+package lani
 
 import "errors"
