+++ pkg/store/lani/mock.go
@@ -1,4 +1,4 @@
-package store
+package lani
 
 import "encoding/binary"
