+++ pkg/store/metadata/metadata.go
@@ -1,18 +1,18 @@
-package store
+package metadata
 
 import (
 	"encoding/binary"
-	"net"
 	"time"
 
 	"github.com/oklog/ulid/v2"
+	"github.com/rotationalio/honu/pkg/store/lani"
 )
 
 //===========================================================================
-// Object Storage
+// Metadata Storage
 //===========================================================================
 
-type Object struct {
+type Metadata struct {
 	Version      *Version
 	Schema       *SchemaVersion
 	MIME         string
@@ -27,10 +27,12 @@ type Object struct {
 	Flags        uint8
 	Created      time.Time
 	Modified     time.Time
-	Data         []byte
 }
 
-func (o *Object) Size() (s int) {
+var _ lani.Encodable = &Metadata{}
+var _ lani.Decodable = &Metadata{}
+
+func (o *Metadata) Size() (s int) {
 	// Version size + not nil bool
 	s += 1
 	if o.Version != nil {
@@ -81,11 +83,10 @@ func (o *Object) Size() (s int) {
 
 	s += 1                         // Flags
 	s += 2 * binary.MaxVarintLen64 // Created, Modified
-	s += len(o.Data) + binary.MaxVarintLen64
 	return
 }
 
-func (o *Object) Encode(e *Encoder) (n int, err error) {
+func (o *Metadata) Encode(e *lani.Encoder) (n int, err error) {
 	var m int
 	if m, err = e.EncodeStruct(o.Version); err != nil {
 		return n + m, err
@@ -166,15 +167,10 @@ func (o *Object) Encode(e *Encoder) (n int, err error) {
 	}
 	n += m
 
-	if m, err = e.Encode(o.Data); err != nil {
-		return n + m, err
-	}
-	n += m
-
 	return
 }
 
-func (o *Object) Decode(d *Decoder) (err error) {
+func (o *Metadata) Decode(d *lani.Decoder) (err error) {
 	// Setup nested structs
 	o.Version = &Version{}
 	o.Schema = &SchemaVersion{}
@@ -265,278 +261,6 @@ func (o *Object) Decode(d *Decoder) (err error) {
 		return err
 	}
 
-	if o.Data, err = d.Decode(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//===========================================================================
-// Object Version
-//===========================================================================
-
-type Version struct {
-	PID       uint64
-	Version   uint64
-	Region    string
-	Parent    *Version
-	Tombstone bool
-	Created   time.Time
-}
-
-func (o *Version) Size() (s int) {
-	s += 2 * binary.MaxVarintLen64
-	s += len([]byte(o.Region)) + binary.MaxVarintLen64
-
-	if o.Parent != nil {
-		s += o.Parent.Size() + 1 // Add 1 for the not nil bool
-	} else {
-		s += 1 // Add 1 for the nil bool
-	}
-
-	s += 1                     // Tombstone bool
-	s += binary.MaxVarintLen64 // Timestamp int64
-
-	return
-}
-
-func (o *Version) Encode(e *Encoder) (n int, err error) {
-	var m int
-	if m, err = e.EncodeUint64(o.PID); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeUint64(o.Version); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeString(o.Region); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeStruct(o.Parent); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeBool(o.Tombstone); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeTime(o.Created); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	return
-}
-
-func (o *Version) Decode(d *Decoder) (err error) {
-	if o.PID, err = d.DecodeUint64(); err != nil {
-		return err
-	}
-
-	if o.Version, err = d.DecodeUint64(); err != nil {
-		return err
-	}
-
-	if o.Region, err = d.DecodeString(); err != nil {
-		return err
-	}
-
-	var isNil bool
-	o.Parent = &Version{}
-
-	if isNil, err = d.DecodeStruct(o.Parent); err != nil {
-		return err
-	}
-
-	if isNil {
-		o.Parent = nil
-	}
-
-	if o.Tombstone, err = d.DecodeBool(); err != nil {
-		return err
-	}
-
-	if o.Created, err = d.DecodeTime(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//===========================================================================
-// Schema Version
-//===========================================================================
-
-type SchemaVersion struct {
-	Name  string
-	Major uint32
-	Minor uint32
-	Patch uint32
-}
-
-func (o *SchemaVersion) Size() (s int) {
-	s += len([]byte(o.Name)) + binary.MaxVarintLen64
-	s += 3 * binary.MaxVarintLen32
-	return
-}
-
-func (o *SchemaVersion) Encode(e *Encoder) (n int, err error) {
-	var m int
-	if m, err = e.EncodeString(o.Name); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeUint32(o.Major); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeUint32(o.Minor); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeUint32(o.Patch); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	return
-}
-
-func (o *SchemaVersion) Decode(d *Decoder) (err error) {
-	if o.Name, err = d.DecodeString(); err != nil {
-		return err
-	}
-
-	if o.Major, err = d.DecodeUint32(); err != nil {
-		return err
-	}
-
-	if o.Minor, err = d.DecodeUint32(); err != nil {
-		return err
-	}
-
-	if o.Patch, err = d.DecodeUint32(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//===========================================================================
-// ACL
-//===========================================================================
-
-type AccessControl struct {
-	ClientID    ulid.ULID
-	Permissions uint8
-}
-
-func (o *AccessControl) Size() int {
-	// ULID + 1 byte
-	return 17
-}
-
-func (o *AccessControl) Encode(e *Encoder) (n int, err error) {
-	var m int
-	if m, err = e.EncodeULID(o.ClientID); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeUint8(o.Permissions); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	return
-}
-
-func (o *AccessControl) Decode(d *Decoder) (err error) {
-	if o.ClientID, err = d.DecodeULID(); err != nil {
-		return err
-	}
-
-	if o.Permissions, err = d.DecodeUint8(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//===========================================================================
-// Provenance
-//===========================================================================
-
-type Publisher struct {
-	PublisherID ulid.ULID
-	ClientID    ulid.ULID
-	IPAddress   net.IP
-	UserAgent   string
-}
-
-func (o *Publisher) Size() (s int) {
-	// 2 ULIDs and 2 variable byte arrays
-	s += 16 + 16
-	s += len(o.IPAddress) + binary.MaxVarintLen64
-	s += len([]byte(o.UserAgent)) + binary.MaxVarintLen64
-	return
-}
-
-func (o *Publisher) Encode(e *Encoder) (n int, err error) {
-	var m int
-	if m, err = e.EncodeULID(o.PublisherID); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeULID(o.ClientID); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.Encode(o.IPAddress); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	if m, err = e.EncodeString(o.UserAgent); err != nil {
-		return n + m, err
-	}
-	n += m
-
-	return
-}
-
-func (o *Publisher) Decode(d *Decoder) (err error) {
-	if o.PublisherID, err = d.DecodeULID(); err != nil {
-		return err
-	}
-
-	if o.ClientID, err = d.DecodeULID(); err != nil {
-		return err
-	}
-
-	var ip []byte
-	if ip, err = d.Decode(); err != nil {
-		return err
-	}
-	o.IPAddress = net.IP(ip)
-
-	if o.UserAgent, err = d.DecodeString(); err != nil {
-		return err
-	}
-
 	return nil
 }
 
@@ -565,6 +289,9 @@ type Encryption struct {
 	SignatureAlgorithm  EncryptionAlgorithm
 }
 
+var _ lani.Encodable = &Encryption{}
+var _ lani.Decodable = &Encryption{}
+
 func (o *Encryption) Size() (s int) {
 	s += len([]byte(o.PublicKeyID)) + binary.MaxVarintLen64
 	s += len(o.EncryptionKey) + binary.MaxVarintLen64
@@ -575,7 +302,7 @@ func (o *Encryption) Size() (s int) {
 	return
 }
 
-func (o *Encryption) Encode(e *Encoder) (n int, err error) {
+func (o *Encryption) Encode(e *lani.Encoder) (n int, err error) {
 	var m int
 	if m, err = e.EncodeString(o.PublicKeyID); err != nil {
 		return n + m, err
@@ -615,7 +342,7 @@ func (o *Encryption) Encode(e *Encoder) (n int, err error) {
 	return
 }
 
-func (o *Encryption) Decode(d *Decoder) (err error) {
+func (o *Encryption) Decode(d *lani.Decoder) (err error) {
 	if o.PublicKeyID, err = d.DecodeString(); err != nil {
 		return err
 	}
@@ -670,11 +397,14 @@ type Compression struct {
 	Level     int64
 }
 
+var _ lani.Encodable = &Compression{}
+var _ lani.Decodable = &Compression{}
+
 func (o *Compression) Size() int {
 	return 1 + binary.MaxVarintLen64
 }
 
-func (o *Compression) Encode(e *Encoder) (n int, err error) {
+func (o *Compression) Encode(e *lani.Encoder) (n int, err error) {
 	var m int
 	if m, err = e.EncodeUint8(uint8(o.Algorithm)); err != nil {
 		return n + m, err
@@ -689,7 +419,7 @@ func (o *Compression) Encode(e *Encoder) (n int, err error) {
 	return
 }
 
-func (o *Compression) Decode(d *Decoder) (err error) {
+func (o *Compression) Decode(d *lani.Decoder) (err error) {
 	var a uint8
 	if a, err = d.DecodeUint8(); err != nil {
 		return err
